Update Fourier_DCT4.c
The 'difference' (upper) half is technically performing a DST-II, but because of the relationship between the DCT and DST, it was implemented 'directly' (ie. reverse and sign-flip every other coefficient). However, this led to an inefficiency in the final 'combine' step where every other coefficient needed to be sign-flipped /again/. So we used the same relationship to avoid reversal in the first stage (which itself is a plus: minus one instruction in SSE, minus two in AVX), which results in not needing to sign-flip in the 'combine' stage. Overall, a pretty good optimization.